72788: Feedback processed: Javadoc, moved service > util, config comment
