+++ dspace/config/modules/usage-statistics.cfg
@@ -42,4 +42,5 @@ usage-statistics.authorization.admin.workflow=true
 #usage-statistics.bots.case-insensitive = false
 
 # Set to true if the statistics core is sharded into a core per year, defaults to false
+# If you are sharding your statistics index each year by running "dspace stats-util -s", you should set this to "true"
 usage-statistics.shardedByYear = false
